+++ test/model.js
@@ -223,3 +223,49 @@ test("persistence", function() {
     post, "destroy", "callback"
   ]);
 });
+
+test("persistence failure", function() {
+  var TestPersistance = {
+    create: function(model, callback) {
+      callback(false);
+    },
+
+    destroy: function(model, callback) {
+      callback(false);
+    },
+
+    update: function(model, callback) {
+      callback(false);
+    }
+  };
+
+  var Post = Model("post", {
+    persistence: TestPersistance
+  });
+
+  var events = [];
+
+  // Stub trigger and capture its argument.
+  Post.prototype.trigger = function(name) {
+    events.push(name);
+  };
+
+  post = new Post();
+  post.save();
+
+  equals(events.join(", "), "initialize",
+    "should not trigger create event if persistence failed");
+
+  same(Post.all(), [], "post should not be added to collection");
+
+  post.attributes.id = 1;
+  post.save();
+
+  equals(events.join(", "), "initialize",
+    "should not trigger update event if persistence failed");
+
+  post.destroy();
+
+  equals(events.join(", "), "initialize",
+    "should not trigger destroy event if persistence failed");
+});